wireless-security-suite/kali-app/requirements.txt
Flask>=2.0
pyyaml
pywifi>=1.3
PyBluez>=0.23
bleak>=0.22.0
pyrtlsdr>=0.7.0
nfcpy>=1.0.3
requests>=2.28
pyttsx3>=2.90
qrcode>=7.4
